Follow imports when going to definitions
This fixes #1033 following the change of behaviour in jedi's
`goto_assignments` in the upgrade from 0.9.0 to 0.11.1. The
change is assumed to be a bug-fix since it now appears to obey
the `follow_imports` boolean for cases which it previously
would follow imports even when the value was set to `False`.

Since we actually do want to follow imports here, we set the
value to `True` to restore the old behaviour.